TASKING F030 RTTD works

+++ srcTrice.C/trice.h
@@ -217,14 +217,14 @@ TRICE_INLINE void triceFifoPush( uint32_t v ){
 //! id trice identifier
 //! \param d16 a 16 bit value
 #define TRICE( id, d16 ) do{ \
-    TRICE_PUSH( (((uint32_t)(d16))<<16) | (id)); \
+    TRICE_PUSH( (((uint32_t)(uint16_t)(d16))<<16) | (id)); \
 } while(0)
 
 //! basic trice macro, assumes d16 to be a 16 bit value
 //! id is 0
 //! \param d16 a 16 bit value
 #define TRICE_ID0( d16 ) do{ \
-    TRICE_PUSH( ((uint32_t)(d16))<<16); \
+    TRICE_PUSH( ((uint32_t)(uint16_t)(d16))<<16); \
 } while(0)
 
 //! trace Id protected (outside critical section)
@@ -253,7 +253,7 @@ TRICE_INLINE void triceFifoPush( uint32_t v ){
 //! \param d1 payload
 #define TRICE8_2( Id, pFmt, d0, d1 ) do{ \
     TRICE_ENTER_CRITICAL_SECTION \
-    TRICE( Id, (((uint32_t)(d1))<<8) | (uint8_t)(d0) ) ; \
+    TRICE( Id,((((uint32_t)((uint8_t)(d1)))<<8) | ((uint8_t)(d0)))) ; \
     TRICE_LEAVE_CRITICAL_SECTION \
 } while(0)
 
@@ -265,7 +265,7 @@ TRICE_INLINE void triceFifoPush( uint32_t v ){
 //! \param d2 payload
 #define TRICE8_3( Id, pFmt, d0, d1, d2 ) do{ \
     TRICE_ENTER_CRITICAL_SECTION \
-    TRICE_ID0( (((uint32_t)(d1))<<8) | (uint8_t)(d0) ) ; \
+    TRICE_ID0( (((uint32_t)((uint8_t)(d1)))<<8) | (uint8_t)(d0) ) ; \
     TRICE( Id, d2 ); \
     TRICE_LEAVE_CRITICAL_SECTION \
 } while(0)
@@ -279,8 +279,8 @@ TRICE_INLINE void triceFifoPush( uint32_t v ){
 //! \param d3 payload
 #define TRICE8_4( Id, pFmt, d0, d1, d2, d3 ) do{ \
     TRICE_ENTER_CRITICAL_SECTION \
-    TRICE_ID0( (((uint32_t)(d1))<<8) | (uint8_t)(d0) ) ; \
-    TRICE( Id, (((uint32_t)(d3))<<8) | (uint8_t)(d2) ) ; \
+    TRICE_ID0(((((uint32_t)(uint8_t)(d1))<<8) | (uint8_t)(d0))) ; \
+    TRICE( Id,((((uint32_t)(uint8_t)(d3))<<8) | (uint8_t)(d2))) ; \
     TRICE_LEAVE_CRITICAL_SECTION \
 } while(0)
 
@@ -294,8 +294,8 @@ TRICE_INLINE void triceFifoPush( uint32_t v ){
 //! \param d4 payload
 #define TRICE8_5( Id, pFmt, d0, d1, d2, d3, d4 ) do{ \
     TRICE_ENTER_CRITICAL_SECTION \
-    TRICE_ID0( (((uint32_t)(d1))<<8) | (uint8_t)(d0) ) ; \
-    TRICE_ID0( (((uint32_t)(d3))<<8) | (uint8_t)(d2) ) ; \
+    TRICE_ID0(((((uint32_t)(uint8_t)(d1))<<8) | (uint8_t)(d0))) ; \
+    TRICE_ID0(((((uint32_t)(uint8_t)(d3))<<8) | (uint8_t)(d2))) ; \
     TRICE( Id, d4 ); \
     TRICE_LEAVE_CRITICAL_SECTION \
 } while(0)
@@ -311,9 +311,9 @@ TRICE_INLINE void triceFifoPush( uint32_t v ){
 //! \param d5 payload
 #define TRICE8_6( Id, pFmt, d0, d1, d2, d3, d4, d5 ) do{ \
     TRICE_ENTER_CRITICAL_SECTION \
-    TRICE_ID0( (((uint32_t)(d1))<<8) | (uint8_t)(d0) ) ; \
-    TRICE_ID0( (((uint32_t)(d3))<<8) | (uint8_t)(d2) ) ; \
-    TRICE( Id, (((uint32_t)(d5))<<8) | (uint8_t)(d4) ) ; \
+    TRICE_ID0(((((uint32_t)(uint8_t)(d1))<<8) | (uint8_t)(d0))) ; \
+    TRICE_ID0(((((uint32_t)(uint8_t)(d3))<<8) | (uint8_t)(d2))) ; \
+    TRICE( Id,((((uint32_t)(uint8_t)(d5))<<8) | (uint8_t)(d4))) ; \
     TRICE_LEAVE_CRITICAL_SECTION \
 } while(0)
 
@@ -329,9 +329,9 @@ TRICE_INLINE void triceFifoPush( uint32_t v ){
 //! \param d6 payload
 #define TRICE8_7( Id, pFmt, d0, d1, d2, d3, d4, d5, d6 ) do{ \
     TRICE_ENTER_CRITICAL_SECTION \
-    TRICE_ID0( (((uint32_t)(d1))<<8) | (uint8_t)(d0) ) ; \
-    TRICE_ID0( (((uint32_t)(d3))<<8) | (uint8_t)(d2) ) ; \
-    TRICE_ID0( (((uint32_t)(d5))<<8) | (uint8_t)(d4) ) ; \
+    TRICE_ID0( (((uint32_t)(uint8_t)(d1))<<8) | (uint8_t)(d0) ) ; \
+    TRICE_ID0( (((uint32_t)(uint8_t)(d3))<<8) | (uint8_t)(d2) ) ; \
+    TRICE_ID0( (((uint32_t)(uint8_t)(d5))<<8) | (uint8_t)(d4) ) ; \
     TRICE( Id, d6 ); \
     TRICE_LEAVE_CRITICAL_SECTION \
 } while(0)
@@ -349,10 +349,10 @@ TRICE_INLINE void triceFifoPush( uint32_t v ){
 //! \param d7 payload
 #define TRICE8_8( Id, pFmt, d0, d1, d2, d3, d4, d5, d6, d7 ) do{ \
     TRICE_ENTER_CRITICAL_SECTION \
-    TRICE_ID0( (((uint32_t)(d1))<<8) | (uint8_t)(d0) ) ; \
-    TRICE_ID0( (((uint32_t)(d3))<<8) | (uint8_t)(d2) ) ; \
-    TRICE_ID0( (((uint32_t)(d5))<<8) | (uint8_t)(d4) ) ; \
-    TRICE( Id, (((uint32_t)(d7))<<8) | (uint8_t)(d6) ) ; \
+    TRICE_ID0(((((uint32_t)(uint8_t)(d1))<<8) | (uint8_t)(d0))) ; \
+    TRICE_ID0(((((uint32_t)(uint8_t)(d3))<<8) | (uint8_t)(d2))) ; \
+    TRICE_ID0(((((uint32_t)(uint8_t)(d5))<<8) | (uint8_t)(d4))) ; \
+    TRICE( Id,((((uint32_t)(uint8_t)(d7))<<8) | (uint8_t)(d6))) ; \
     TRICE_LEAVE_CRITICAL_SECTION \
 } while(0)
